% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{areas}
\alias{areas}
\title{Areas of the Polygons of a SpatialPolygon* Object}
\usage{
areas(spplgn)
}
\description{
Returns a list of the areas of the Polygon objects of the polygons slot of
a SpatialPolygon* object.
}
\details{
The returned list is of same length as the input SpatialPolygon* object. Each
slot of the list is a numeric vector, the lenght of which is equal to the
number of Polygon objects constituing
}
