% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marc_datafield.R
\name{as.data.frame.marcdatafield}
\alias{as.data.frame.marcdatafield}
\title{Simplifie Marc Datafield to a Data.frame}
\usage{
\method{as.data.frame}{marcdatafield}(x, ...)
}
\arguments{
\item{x}{Marcdatafield object to coerce into a data.frame}

\item{...}{Additional arguments. Will be ignored.}
}
\value{
All values from the provided object in a data.frame, one row per code-value pair
}
\description{
Simplifie Marc Datafield to a Data.frame
}
