% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdlist.R
\name{birdlist}
\alias{birdlist}
\title{Print bird list from eBird report}
\usage{
birdlist(file = "~/Desktop/eBirdReports.xls")
}
\description{
The eBird report should be generated from the
"My eBird" > "Summarize My Observations" and then "Download this report".
The report in question should be a tabulation-separated text file named
"eBirdReports.xls".
}
\author{
Marc Choisy
}
