% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marc_datafield.R
\name{get_field_values}
\alias{get_field_values}
\title{Get Values from Marc Datafield}
\usage{
get_field_values(.data, codes, include_code = FALSE)
}
\arguments{
\item{.data}{A marcdatafield object or a list of marcdatafield objects}

\item{codes}{One or multiple codes for which to retrieve the values}

\item{include_code}{If true the returned values will be named with their original code}
}
\value{
Either a vector of the matched values or a list of those vectors, depending on input. NA if no codes were matched.
}
\description{
Get Values from Marc Datafield
}
