% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll2incidence.R
\name{ll2incidence}
\alias{ll2incidence}
\title{Calculates incidence from line listing}
\usage{
ll2incidence(x, unit = c("day", "week", "month"))
}
\arguments{
\item{\code{x}}{a vector of dates of cases}

\item{\code{unit}}{a character string ("day", "week", "month") specifying the
temporal aggregation wished for the incidence calculation.}
}
\description{
Calculates daily, weekly or monthly incidence from line listing.
}
