% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_film_list.R
\name{make_film_list}
\alias{make_film_list}
\title{Write the films list}
\usage{
make_film_list(from = "/Volumes/Seagate Backup Plus Drive/films",
  to = "~/Dropbox/Public/film_list.html", ftp = "ftpperso.free.fr",
  login = "marcchoisy", passwd)
}
\description{
Write the films list
}
\author{
Marc Choisy
}
