% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marc_record.R
\name{MarcRecord}
\alias{MarcRecord}
\title{R6 Class Representing a Marc21 record}
\description{
This is the base class used to interact with marc records
}
\details{
This class provides functions for reading marc records from online or disk sources
and to retrieve information from them.

Currently only  MarcXML files are supported.
}
\examples{

## ------------------------------------------------
## Method `MarcRecord$get_fields`
## ------------------------------------------------

record <- MarcRecord$new()
record$read_record("http://d-nb.info/gnd/11897792X/about/marcxml")

# Get fields with optional filtering
record$get_fields(tag = c(548, 550), simplify = TRUE)

## ------------------------------------------------
## Method `MarcRecord$read_record`
## ------------------------------------------------

record <- MarcRecord$new()
record$read_record("http://d-nb.info/gnd/11897792X/about/marcxml")
record
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{controlfields}}{The control fields (00X) of the record}

\item{\code{datafields}}{The data fields (>= 010) of the record}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_fields}{\code{MarcRecord$get_fields()}}
\item \href{#method-read_record}{\code{MarcRecord$read_record()}}
\item \href{#method-print}{\code{MarcRecord$print()}}
\item \href{#method-clone}{\code{MarcRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_fields}{}}}
\subsection{Method \code{get_fields()}}{
Retrieve data fields from this record while applying optional filtering.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarcRecord$get_fields(tag, ind_1, ind_2, simplify = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag}}{Optional. Tag(s) for which to filter the record, must be a vector of size >= 1 and type numeric or character}

\item{\code{ind_1}}{Optional. First indicator(s) for which to filter the record, must be a vector of size >= 1 and type character}

\item{\code{ind_2}}{Optional. Second indicator(s) for which to filter the record, must be a vector of size >= 1 and type character}

\item{\code{simplify}}{If set to true the data fields will be coerced in to a data frame before returning, otherwise a list of \code{\link[maRc]{MarcDatafield}} will be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The requested data as either a data frame or a list, depending on \code{simplify}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{record <- MarcRecord$new()
record$read_record("http://d-nb.info/gnd/11897792X/about/marcxml")

# Get fields with optional filtering
record$get_fields(tag = c(548, 550), simplify = TRUE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_record"></a>}}
\if{latex}{\out{\hypertarget{method-read_record}{}}}
\subsection{Method \code{read_record()}}{
Read a marc record from disk or an online source.
Currently only marcxml is supported.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarcRecord$read_record(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path or URL of the data to read. Must be a character vector of size 1.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{record <- MarcRecord$new()
record$read_record("http://d-nb.info/gnd/11897792X/about/marcxml")
record
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print method for a marc record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarcRecord$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarcRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
