% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovv.R.R
\name{ovv}
\alias{ovv}
\title{Overview of a data frame.}
\usage{
ovv(df, digits = 4, interspace = 3)
}
\arguments{
\item{x}{A vector of characters.}
}
\value{
The translation of x in Basic Latin.
}
\description{
\code{ovv} returns an overview of a data frame.
}
\examples{
2+2
}
\author{
Marc Choisy
}

