% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marc_record.R
\name{read_marcxml}
\alias{read_marcxml}
\title{Read MarcXML files}
\usage{
read_marcxml(source)
}
\arguments{
\item{source}{File to read. Can be an url, a path or literal xml.}
}
\value{
Parsed record as marcrecord object
}
\description{
This function will read a MarcXML file and parse it into an marcrecord object.
}
