% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{send}
\alias{send}
\title{Send an R object by email}
\usage{
send(object, to, body, subject)
}
\arguments{
\item{x}{A vector of characters.}
}
\value{
The translation of x in Basic Latin.
}
\description{
This function sends an R object by email in the form of an RDS object. Email
address can be specified explicitly or defined by a pseudo from the addresses
book (see function \code{addresses} for managing the addresses book). To be
able to use this functionn the user needs to do some configuration first that
include:
}
\details{
1. we recommand the user creates a gmail email account that will be used
exclusively by the \code{send} function. It's safer in terms of security than
using the user's normal email account.

2. if you are using Mac OS X, configure your system following the steps from
here URL
}
\examples{
2+2
}
\seealso{
The function \code{addresses} allows to manage the addresses book and
see the pseudo that can be used in the \code{send} function in place of a
fully defined email address.
}
\author{
Marc Choisy
}
