% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendRDS.R
\name{sendRDS}
\alias{sendRDS}
\title{Send an R object by email.}
\usage{
sendRDS(object, to)
}
\arguments{
\item{x}{A vector of characters.}
}
\value{
The translation of x in Basic Latin.
}
\description{
\code{ovv} returns an overview of a data frame.
}
\examples{
2+2
}
\author{
Marc Choisy
}

