% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_disp.R
\name{test_disp}
\alias{test_disp}
\title{Test of dispersion}
\source{
method suggested in the book "R in action" by Kabacoff (2011).
}
\usage{
test_disp(m)
}
\arguments{
\item{m}{a glm object with family equal to "poisson" of "binomial".}
}
\value{
returns a named vector of 4 components: dispersion parameter ("phi"),
statistic ("stat"), number of degrees of freedom ("df") and corresponding p
value from the test ("p").
}
\description{
\code{test_disp} tests the significativity of the dispersion in a Poisson
or logistic model.
}
\examples{
m <- glm(Species=="versicolor" ~ Sepal.Width,family=binomial(), data=iris)
test_disp(m)
}
\author{
Marc Choisy
}
