% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_base_latin.R
\name{to_basic_latin}
\alias{to_basic_latin}
\title{Conversion to base Latin.}
\source{
\url{https://en.wikipedia.org/wiki/Latin-1_Supplement_(Unicode_block)},
\url{https://en.wikipedia.org/wiki/Latin_Extended-A},
\url{https://en.wikipedia.org/wiki/Latin_Extended-B} and
\url{https://en.wikipedia.org/wiki/Latin_Extended_Additional}.
}
\usage{
to_basic_latin(x)
}
\arguments{
\item{x}{A vector of characters.}
}
\value{
The translation of x in Basic Latin.
}
\description{
\code{to_base_latin} returns the base Latin equivalent of non-base Latin
Unicode characters.
}
\details{
As input could be any Unicode character of the Latin-1 Supplement, Latin
Extended-A & B, and Latin Extended Additional blocks of the Unicode standard.
The output will be the equivalent in the Basic Latin block.
}
\examples{
to_basic_latin(c("H<U+00E0> Noi","Hai ph<U+00F2>ng","H<U+00E0> Giang"))
}

